Added methods for ingredientapplicationimpl and ingredient controller

+++ src/main/java/com/pizzeria/domain/ingredientdomain/IngredientProjection.java
@@ -1,11 +1,12 @@
 package com.pizzeria.domain.ingredientdomain;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface IngredientProjection {
 	
 	public UUID getId();
 	public String getName();
-	public Double getPrice();
+	public BigDecimal getPrice();
 	
 }
